input switch for light movement, new file for cylinder hit and norm for cylinder hit
